Refactor

+++ lib/preprocessor.py
@@ -2,9 +2,8 @@ import os
 import xml.etree.ElementTree as ET
 import json
 import numpy as np
-from enum import Enum
-from lib.step import IStep
-from lib.util.path import get_path_gml, get_path_json
+from lib.util.path import PathUtil
+from lib.util.lod import Level
 
 CORE = "{http://www.opengis.net/citygml/2.0}"
 GML = "{http://www.opengis.net/gml}"
@@ -12,36 +11,36 @@ GEN = "{http://www.opengis.net/citygml/generics/2.0}"
 ID = "{http://www.opengis.net/gml}id"
 
 
-class Level(Enum):
-    LOD1 = 1
-    LOD2 = 2
+class Preprocessor():
+
+    def process(self, level):
+        print("Parsing GML files and extracting useful info...")
+
+        path_util = PathUtil(level)
+        data_dir_path = path_util.get_data_dir_path()
+
+        for filename in os.listdir(data_dir_path):
+            file_path = os.path.join(data_dir_path, filename)
+
+            if os.path.isfile(file_path) and file_path.endswith(".gml"):
+                print(f'Processing {filename}')
+
+                tree = ET.parse(file_path)
+                root = tree.getroot()
+
+                attribute_map = {}
+                buildings = root.findall(f"{CORE}cityObjectMember")
+                self.__process_buildings(buildings, attribute_map, level)
+
+                json_name = filename.replace(".gml", "")
+
+                with open(path_util.get_path_json(json_name), 'w') as fp:
+                    json.dump(attribute_map, fp)
+            else:
+                print(f'Ignoring file: {file_path}')
 
 
-class Preprocessor(IStep):
-    def __init__(self, file_gdrive_id, level):
-        self.__file_gdrive_id = file_gdrive_id
-        self.__level = level
-
-    def execute(self):
-        print("Parsing GML and extracting useful info...")
-
-        path = get_path_gml(self.__file_gdrive_id)
-        tree = ET.parse(path)
-        root = tree.getroot()
-
-        attribute_map = {}
-        buildings = root.findall(f"{CORE}cityObjectMember")
-        self.__process_buildings(buildings, attribute_map)
-
-        with open(get_path_json(self.__file_gdrive_id), 'w') as fp:
-            json.dump(attribute_map, fp)
-
-
-    def cleanup(self):
-        os.remove(get_path_gml(self.__file_gdrive_id))
-
-
-    def __process_buildings(self, buildings, attribute_map):
+    def __process_buildings(self, buildings, attribute_map, level):
         for building in buildings:
             id = building[0].attrib[ID]
             roof_height = self.__get_roof_height(building)
@@ -51,8 +50,8 @@ class Preprocessor(IStep):
             for points_set in building.iter(f"{GML}posList"):
                 surface = Surface(points_set.text)
 
-                should_process_lod1 = self.__level == Level.LOD1 and surface.is_lod1_roof(roof_height)
-                should_process_lod2 = self.__level == Level.LOD2 and surface.is_lod2_roof(roof_height)
+                should_process_lod1 = level == Level.LOD1 and surface.is_lod1_roof(roof_height)
+                should_process_lod2 = level == Level.LOD2 and surface.is_lod2_roof(roof_height)
 
                 if should_process_lod1 or should_process_lod2:
                     area = surface.area()
@@ -92,7 +91,8 @@ class Surface:
     def is_lod1_roof(self, roof_height) -> bool:
         is_roof = True
         for point in self.points:
-            if point.z != roof_height:
+            z = point[2]
+            if z != roof_height:
                 is_roof = False
                 break
         
@@ -104,10 +104,14 @@ class Surface:
 
     # https://stackoverflow.com/questions/12642256/find-area-of-polygon-from-xyz-coordinates
     def area(self):
+        poly = np.array(self.points)
         #all edges
-        edges = self.points[1:] - self.points[0:1]
+        edges = poly[1:] - poly[0:1]
         # row wise cross product
         cross_product = np.cross(edges[:-1],edges[1:], axis=1)
         #area of all triangles
         area = np.linalg.norm(cross_product, axis=1) / 2
         return sum(area)
+
+    def incline(self):
+        return 0.0
